test: move hbm_pool e2e tests from test_coarse_tile_e2e.py
Relocate the five hbm_pool/pool_size threading tests
(test_bundle_pool_size_threaded_from_hbm_pool_sizes,
test_pool_alloc_scoped_per_bundle_across_fallback_boundary,
test_no_python_side_pool_tensor_allocated,
test_pool_size_kwarg_in_generated_sdsc_call,
test_pool_size_kwarg_omitted_when_no_pool) into a new
TestHbmPoolPlanningE2E class in test_hbm_pool_planning.py, alongside the
existing hbm_pool_planning unit tests, rather than in the coarse-tile e2e
file.

Fix a latent bug the move exposed: two tests asserted assertNotIn("_pool_",
src) to check for the absence of a Python-side pool tensor, but this
file's own name (test_hbm_pool_planning.py) contains "_pool_" and leaks
into generated source via embedded SourceLoc debug-handle file paths.
Replaced with a regex matching the actual _pool_{kernel_name} = ...
assignment pattern emitted by call_kernel().
